app/requirements.txt
discord.py>=2.3.0
anthropic>=0.34.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
redis>=5.0.0
hiredis>=2.3.0
watchdog>=3.0.0
cryptography>=41.0.0